+++ shared/src/lib.rs
@@ -89,10 +89,11 @@ pub use messages::{
         channel::{Channel, ChannelDirection, ChannelMode, ReliableSettings, TickBufferSettings},
         channel_kinds::{ChannelKind, ChannelKinds},
         default_channels,
-        message_channel::{ChannelReceiver, ChannelSender},
-        ordered_reliable_receiver::OrderedReliableReceiver,
-        reliable_sender::ReliableSender,
-        unordered_reliable_receiver::UnorderedReliableReceiver,
+        receivers::{
+            channel_receiver::ChannelReceiver, ordered_reliable_receiver::OrderedReliableReceiver,
+            unordered_reliable_receiver::UnorderedReliableReceiver,
+        },
+        senders::{channel_sender::ChannelSender, reliable_sender::ReliableSender},
     },
     message::{Message, Message as MessageBevy, Message as MessageHecs, MessageBuilder},
     message_container::MessageContainer,
