+++ shared/src/messages/channels/mod.rs
@@ -1,17 +1,5 @@
 pub mod channel;
 pub mod channel_kinds;
 pub mod default_channels;
-pub mod indexed_message_reader;
-pub mod indexed_message_writer;
-pub mod message_channel;
-pub mod ordered_reliable_receiver;
-pub mod reliable_sender;
-pub mod sequenced_reliable_receiver;
-pub mod sequenced_unreliable_receiver;
-pub mod sequenced_unreliable_sender;
-pub mod unordered_reliable_receiver;
-pub mod unordered_unreliable_receiver;
-pub mod unordered_unreliable_sender;
-
-pub mod fragment_receiver;
-mod reliable_receiver;
+pub mod receivers;
+pub mod senders;
