+++ shared/src/messages/channels/receivers/indexed_message_reader.rs
