+++ shared/src/messages/channels/receivers/unordered_reliable_receiver.rs
@@ -1,5 +1,5 @@
 use crate::{
-    messages::channels::reliable_receiver::{ReceiverArranger, ReliableReceiver},
+    messages::channels::receivers::reliable_receiver::{ReceiverArranger, ReliableReceiver},
     types::MessageIndex,
     MessageContainer,
 };
