+++ shared/src/messages/channels/senders/channel_sender.rs
@@ -1,4 +1,4 @@
-use naia_serde::{BitReader, BitWriter, SerdeErr};
+use naia_serde::BitWriter;
 use naia_socket_shared::Instant;
 
 use crate::{
@@ -27,19 +27,4 @@ pub trait MessageChannelSender: ChannelSender<MessageContainer> {
         writer: &mut BitWriter,
         has_written: &mut bool,
     ) -> Option<Vec<MessageIndex>>;
-}
-
-pub trait ChannelReceiver<P>: Send + Sync {
-    /// Read messages from an internal buffer and return their content
-    fn receive_messages(&mut self) -> Vec<P>;
-}
-
-pub trait MessageChannelReceiver: ChannelReceiver<MessageContainer> {
-    /// Read messages from raw bits, parse them and store then into an internal buffer
-    fn read_messages(
-        &mut self,
-        message_kinds: &MessageKinds,
-        converter: &dyn NetEntityHandleConverter,
-        reader: &mut BitReader,
-    ) -> Result<(), SerdeErr>;
 }
