+++ shared/src/messages/channels/senders/indexed_message_writer.rs
