+++ shared/src/messages/mod.rs
@@ -1,7 +1,7 @@
 pub mod channels;
+pub mod fragment;
 pub mod message;
 pub mod message_container;
-pub mod message_fragmenter;
 pub mod message_kinds;
 pub mod message_manager;
 pub mod named;
